Move type to Infra
