+++ x-pack/plugins/infra/public/pages/logs/settings/log_columns_configuration_panel.tsx
@@ -22,7 +22,7 @@ import {
 import { i18n } from '@kbn/i18n';
 import { FormattedMessage } from '@kbn/i18n-react';
 import React, { useCallback } from 'react';
-import { DragHandleProps, DropResult } from '@kbn/observability-plugin/public';
+import type { DropResult, DragHandleProps } from '../../../types';
 import {
   FieldLogColumnConfiguration,
   getLogColumnConfigurationId,
