+++ x-pack/plugins/observability/public/typings/index.ts
@@ -5,7 +5,6 @@
  * 2.0.
  */
 
-export * from './eui_draggable';
 export * from './fetch_overview_data';
 export * from './slo';
 export * from './utils';
